chore: Move useCollaborationContext to dedicated file
