+++ .flowconfig
@@ -76,6 +76,7 @@ module.name_mapper='^@lexical/react/LexicalTableOfContents__EXPERIMENTAL' -> '<P
 module.name_mapper='^@lexical/react/LexicalAutoFormatterPlugin' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalAutoFormatterPlugin.js.flow'
 module.name_mapper='^@lexical/react/LexicalCharacterLimitPlugin' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalCharacterLimitPlugin.js.flow'
 module.name_mapper='^@lexical/react/LexicalClearEditorPlugin' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalClearEditorPlugin.js.flow'
+module.name_mapper='^@lexical/react/LexicalCollaborationContext' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalCollaborationContext.js.flow'
 module.name_mapper='^@lexical/react/LexicalCollaborationPlugin' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalCollaborationPlugin.js.flow'
 module.name_mapper='^@lexical/react/LexicalTablePlugin' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalTablePlugin.js.flow'
 module.name_mapper='^@lexical/react/LexicalLinkPlugin' -> '<PROJECT_ROOT>/packages/lexical-react/flow/LexicalLinkPlugin.js.flow'
