+++ packages/lexical-react/src/__tests__/unit/utils.tsx
@@ -13,10 +13,8 @@ import {createRoot, Root} from 'react-dom/client';
 import * as ReactTestUtils from 'react-dom/test-utils';
 import * as Y from 'yjs';
 
-import {
-  CollaborationPlugin,
-  useCollaborationContext,
-} from '../../LexicalCollaborationPlugin';
+import {useCollaborationContext} from '../../LexicalCollaborationContext';
+import {CollaborationPlugin} from '../../LexicalCollaborationPlugin';
 import {LexicalComposer} from '../../LexicalComposer';
 import {ContentEditable} from '../../LexicalContentEditable';
 import {RichTextPlugin} from '../../LexicalRichTextPlugin';
