+++ tsconfig.json
@@ -135,6 +135,9 @@
       "@lexical/react/LexicalClearEditorPlugin": [
         "./packages/lexical-react/src/LexicalClearEditorPlugin.ts"
       ],
+      "@lexical/react/LexicalCollaborationContext": [
+        "./packages/lexical-react/src/LexicalCollaborationContext.ts"
+      ],
       "@lexical/react/LexicalCollaborationPlugin": [
         "./packages/lexical-react/src/LexicalCollaborationPlugin.ts"
       ],
